add default for db.updateJobs adapter
