+++ packages/payload/src/database/createDatabaseAdapter.ts
@@ -8,6 +8,7 @@ import type {
   RollbackTransaction,
 } from './types.js'
 
+import { defaultUpdateJobs } from './defaultUpdateJobs.js'
 import { createMigration } from './migrations/createMigration.js'
 import { migrate } from './migrations/migrate.js'
 import { migrateDown } from './migrations/migrateDown.js'
@@ -31,6 +32,7 @@ export function createDatabaseAdapter<T extends BaseDatabaseAdapter>(
     | 'migrateReset'
     | 'migrateStatus'
     | 'migrationDir'
+    | 'updateJobs'
   >,
 ): T {
   return {
@@ -45,6 +47,7 @@ export function createDatabaseAdapter<T extends BaseDatabaseAdapter>(
     migrateReset,
     migrateStatus,
     rollbackTransaction,
+    updateJobs: defaultUpdateJobs,
 
     ...args,
     // Ensure migrationDir is set
